refactor: consolidate service imports for improved maintainability and clarity added for ui elemnts to use one service barrel file for the imports of the services
