+++ bmusic-ui/src/app/home/albums/album/album.component.ts
@@ -1,13 +1,14 @@
-import { Component, OnInit, OnDestroy } from '@angular/core';
-import { SongsStateService } from '../../../service/states/songs.state.service';
-import { Song } from '../../../service/models/song.interface';
+import { Component, OnInit } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { AlbumStateService } from '../../../service/states/album.state.service';
-import { ApiService } from '../../../service/apiCalls/api.service';
 import { ActivatedRoute } from '@angular/router';
 import { SongsListComponent } from '../../shared/songs-list/songs-list.component';
-import { SortService } from '../../../service/states/shared/sort.service';
-import { environment } from '../../../../environments/environment';
+import {
+  SortService,
+  ApiService,
+  AlbumStateService,
+  Song,
+  SongsStateService,
+} from '../../../service';
 
 @Component({
   selector: 'app-album',
