+++ bmusic-ui/src/app/home/artists/artist/artist.component.ts
@@ -1,13 +1,15 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 import { CommonModule } from '@angular/common';
-import { SongsStateService } from '../../../service/states/songs.state.service';
-import { ArtistStateService } from '../../../service/states/artist.state.service';
-import { Song } from '../../../service/models/song.interface';
-import { ApiService } from '../../../service/apiCalls/api.service';
-import { AlbumStateService } from '../../../service/states/album.state.service';
 import { SongsListComponent } from '../../shared/songs-list/songs-list.component';
-import { SortService } from '../../../service/states/shared/sort.service';
+import {
+  SortService,
+  AlbumStateService,
+  ApiService,
+  Song,
+  ArtistStateService,
+  SongsStateService,
+} from '../../../service';
 
 @Component({
   selector: 'app-artist',
