+++ bmusic-ui/src/app/home/playlists/playlist/playlist.component.ts
@@ -2,10 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 import { CommonModule } from '@angular/common';
 import { SongsListComponent } from '../../shared/songs-list/songs-list.component';
-import { SongsStateService } from '../../../service/states/songs.state.service';
-import { ApiPlaylistService } from '../../../service/apiCalls/api-playlist.service';
-import { Playlist } from '../../../service/models/playlist.interface';
-import { Song } from '../../../service/models/song.interface';
+import { Song, Playlist, ApiPlaylistService } from '../../../service';
 import { Observable } from 'rxjs';
 
 @Component({
