+++ bmusic-ui/src/app/home/songs/songs.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit, ChangeDetectionStrategy } from '@angular/core';
+import { Component, ChangeDetectionStrategy } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { SongsListComponent } from '../shared/songs-list/songs-list.component';
